Rework messages logic as unity-like

+++ PeacefulHills/Assets/Scripts/Shared/Network/Messages/MessageInfo.cs
@@ -5,9 +5,9 @@ namespace PeacefulHills.Network.Messages
     public readonly struct MessageInfo
     {
         public readonly TypeManager.TypeInfo TypeInfo;
-        public readonly uint Id;
+        public readonly ushort Id;
 
-        public MessageInfo(TypeManager.TypeInfo typeInfo, uint id)
+        public MessageInfo(TypeManager.TypeInfo typeInfo, ushort id)
         {
             TypeInfo = typeInfo;
             Id = id;
